Change client page title from ISPyB to SynchWeb
While it's true that the underlying database schema is ISPyB, what the
user really sees and uses is SynchWeb.  The API is a SynchWeb API, so
SynchWeb is what a developer sees and uses as well.  To make this more
clear, change ISPyB in the client page title to SynchWeb.